+++ resources/sql/queries.sql
@@ -91,3 +91,15 @@ INSERT INTO points
 (from_user, to_user, pt)
 VALUES
 (:from, :to, :pt)
+
+-- :name points-from :? :*
+-- :doc info about `login` sent points.
+select pt, count(pt) from points
+where from_user = :login
+group by pt;
+
+-- :name points-to :? :*
+-- :doc info about `login` received points.
+select pt, count(pt) from points
+where to_user = :login
+group by pt;